requirements.txt
python-telegram-bot==20.3
openai==1.0
sentry-sdk==1.0
prometheus-client==0.17
python-dotenv==0.19